requirements.txt
twilio
flask
pandas
newsapi-python